Synchro Repo

+++ main.go
@@ -3,13 +3,28 @@ package main
 import (
 	"bufio"
 	"bytes"
+	"context"
+	"encoding/json"
+	"io/ioutil"
+	"net/http"
 	"os"
+	"path/filepath"
 	"regexp"
 	"strings"
+	"time"
 
 	"github.com/eth0izzle/shhgit/core"
 	"github.com/fatih/color"
 )
+
+type MatchEvent struct {
+	Url       string
+	Matches   []string
+	Signature string
+	File      string
+	Stars     int
+	Source    core.GitResourceType
+}
 
 var session = core.GetSession()
 
@@ -18,13 +33,17 @@ func ProcessRepositories() {
 
 	for i := 0; i < threadNum; i++ {
 		go func(tid int) {
-
 			for {
-				repositoryId := <-session.Repositories
-				repo, err := core.GetRepository(session, repositoryId)
+				timeout := time.Duration(*session.Options.CloneRepositoryTimeout) * time.Second
+				_, cancel := context.WithTimeout(context.Background(), timeout)
+				defer cancel()
+
+				repository := <-session.Repositories
+
+				repo, err := core.GetRepository(session, repository.Id)
 
 				if err != nil {
-					session.Log.Warn("Failed to retrieve repository %d: %s", repositoryId, err)
+					session.Log.Warn("Failed to retrieve repository %d: %s", repository.Id, err)
 					continue
 				}
 
@@ -32,7 +51,7 @@ func ProcessRepositories() {
 					uint(repo.GetStargazersCount()) >= *session.Options.MinimumStars &&
 					uint(repo.GetSize()) < *session.Options.MaximumRepositorySize {
 
-					processRepositoryOrGist(repo.GetCloneURL())
+					processRepositoryOrGist(repo.GetCloneURL(), repository.Ref, repo.GetStargazersCount(), core.GITHUB_SOURCE)
 				}
 			}
 		}(i)
@@ -46,20 +65,37 @@ func ProcessGists() {
 		go func(tid int) {
 			for {
 				gistUrl := <-session.Gists
-				processRepositoryOrGist(gistUrl)
+				processRepositoryOrGist(gistUrl, "", -1, core.GIST_SOURCE)
 			}
 		}(i)
 	}
 }
 
-func processRepositoryOrGist(url string) {
+func ProcessComments() {
+	threadNum := *session.Options.Threads
+
+	for i := 0; i < threadNum; i++ {
+		go func(tid int) {
+			for {
+				commentBody := <-session.Comments
+				dir := core.GetTempDir(core.GetHash(commentBody))
+				ioutil.WriteFile(filepath.Join(dir, "comment.ignore"), []byte(commentBody), 0644)
+
+				if !checkSignatures(dir, "ISSUE", 0, core.GITHUB_COMMENT) {
+					os.RemoveAll(dir)
+				}
+			}
+		}(i)
+	}
+}
+
+func processRepositoryOrGist(url string, ref string, stars int, source core.GitResourceType) {
 	var (
-		matches    []string
 		matchedAny bool = false
 	)
 
 	dir := core.GetTempDir(core.GetHash(url))
-	_, err := core.CloneRepository(session, url, dir)
+	_, err := core.CloneRepository(session, url, ref, dir)
 
 	if err != nil {
 		session.Log.Debug("[%s] Cloning failed: %s", url, err.Error())
@@ -67,10 +103,24 @@ func processRepositoryOrGist(url string) {
 		return
 	}
 
-	session.Log.Debug("[%s] Cloning in to %s", url, strings.Replace(dir, *session.Options.TempDirectory, "", -1))
+	session.Log.Debug("[%s] Cloning %s in to %s", url, ref, strings.Replace(dir, *session.Options.TempDirectory, "", -1))
+	matchedAny = checkSignatures(dir, url, stars, source)
+	if !matchedAny {
+		os.RemoveAll(dir)
+	}
+}
 
+func checkSignatures(dir string, url string, stars int, source core.GitResourceType) (matchedAny bool) {
 	for _, file := range core.GetMatchingFiles(dir) {
-		relativeFileName := strings.Replace(file.Path, *session.Options.TempDirectory, "", -1)
+		var (
+			matches          []string
+			relativeFileName string
+		)
+		if strings.Contains(dir, *session.Options.TempDirectory) {
+			relativeFileName = strings.Replace(file.Path, *session.Options.TempDirectory, "", -1)
+		} else {
+			relativeFileName = strings.Replace(file.Path, dir, "", -1)
+		}
 
 		if *session.Options.SearchQuery != "" {
 			queryRegex := regexp.MustCompile(*session.Options.SearchQuery)
@@ -90,14 +140,16 @@ func processRepositoryOrGist(url string) {
 					matchedAny = true
 
 					if part == core.PartContents {
-						if matches = signature.GetContentsMatches(file); matches != nil {
+						if matches = signature.GetContentsMatches(file.Contents); matches != nil {
 							count := len(matches)
 							m := strings.Join(matches, ", ")
+							publish(&MatchEvent{Source: source, Url: url, Matches: matches, Signature: signature.Name(), File: relativeFileName, Stars: stars})
 							session.Log.Important("[%s] %d %s for %s in file %s: %s", url, count, core.Pluralize(count, "match", "matches"), color.GreenString(signature.Name()), relativeFileName, color.YellowString(m))
 							session.WriteToCsv([]string{url, signature.Name(), relativeFileName, m})
 						}
 					} else {
 						if *session.Options.PathChecks {
+							publish(&MatchEvent{Source: source, Url: url, Matches: matches, Signature: signature.Name(), File: relativeFileName, Stars: stars})
 							session.Log.Important("[%s] Matching file %s for %s", url, color.YellowString(relativeFileName), color.GreenString(signature.Name()))
 							session.WriteToCsv([]string{url, signature.Name(), relativeFileName, ""})
 						}
@@ -109,11 +161,22 @@ func processRepositoryOrGist(url string) {
 								line := scanner.Text()
 
 								if len(line) > 6 && len(line) < 100 {
-									entropy := core.GetEntropy(scanner.Text())
+									entropy := core.GetEntropy(line)
 
 									if entropy >= *session.Options.EntropyThreshold {
-										session.Log.Important("[%s] Potential secret in %s = %s", url, color.YellowString(relativeFileName), color.GreenString(scanner.Text()))
-										session.WriteToCsv([]string{url, signature.Name(), relativeFileName, scanner.Text()})
+										blacklistedMatch := false
+
+										for _, blacklistedString := range session.Config.BlacklistedStrings {
+											if strings.Contains(strings.ToLower(line), strings.ToLower(blacklistedString)) {
+												blacklistedMatch = true
+											}
+										}
+
+										if !blacklistedMatch {
+											publish(&MatchEvent{Source: source, Url: url, Matches: []string{line}, Signature: "High entropy string", File: relativeFileName, Stars: stars})
+											session.Log.Important("[%s] Potential secret in %s = %s", url, color.YellowString(relativeFileName), color.GreenString(line))
+											session.WriteToCsv([]string{url, "High entropy string", relativeFileName, line})
+										}
 									}
 								}
 							}
@@ -123,31 +186,51 @@ func processRepositoryOrGist(url string) {
 			}
 		}
 
-		if !matchedAny {
+		if !matchedAny && len(*session.Options.Local) <= 0 {
 			os.Remove(file.Path)
 		}
 	}
+	return
+}
 
-	if !matchedAny {
-		os.RemoveAll(dir)
+func publish(event *MatchEvent) {
+	// todo: implement a modular plugin system to handle the various outputs (console, live, csv, webhooks, etc)
+	if len(*session.Options.Live) > 0 {
+		data, _ := json.Marshal(event)
+		http.Post(*session.Options.Live, "application/json", bytes.NewBuffer(data))
 	}
 }
 
 func main() {
-	session.Log.Info("%s v%s started. Loaded %d signatures. Using %d GitHub tokens and %d threads. Work dir: %s", core.Name, core.Version, len(session.Signatures), len(session.Clients), *session.Options.Threads, *session.Options.TempDirectory)
+	session.Log.Info(color.HiBlueString(core.Banner))
+	session.Log.Info("\t%s\n", color.HiCyanString(core.Author))
+	session.Log.Info("[*] Loaded %s signatures. Using %s worker threads. Temp work dir: %s\n", color.BlueString("%d", len(session.Signatures)), color.BlueString("%d", *session.Options.Threads), color.BlueString(*session.Options.TempDirectory))
 
-	if *session.Options.SearchQuery != "" {
-		session.Log.Important("Search Query '%s' given. Only returning matching results.", *session.Options.SearchQuery)
+	if len(*session.Options.Local) > 0 {
+		session.Log.Info("[*] Scanning local directory: %s - skipping public repository checks...", color.BlueString(*session.Options.Local))
+		rc := 0
+		if checkSignatures(*session.Options.Local, *session.Options.Local, -1, core.LOCAL_SOURCE) {
+			rc = 1
+		} else {
+			session.Log.Info("[*] No matching secrets found in %s!", color.BlueString(*session.Options.Local))
+		}
+		os.Exit(rc)
+	} else {
+		if *session.Options.SearchQuery != "" {
+			session.Log.Important("Search Query '%s' given. Only returning matching results.", *session.Options.SearchQuery)
+		}
+
+		go core.GetRepositories(session)
+		go ProcessRepositories()
+		go ProcessComments()
+
+		if *session.Options.ProcessGists {
+			go core.GetGists(session)
+			go ProcessGists()
+		}
+
+		spinny := core.ShowSpinner()
+		select {}
+		spinny()
 	}
-
-	go core.GetRepositories(session)
-	go ProcessRepositories()
-
-	if *session.Options.ProcessGists {
-		go core.GetGists(session)
-		go ProcessGists()
-	}
-
-	session.Log.Info("Press Ctrl+C to stop and exit.\n")
-	select {}
 }
